+++ Player/Scan/CustomScanNodeProperties.cs
@@ -7,6 +7,8 @@ public class CustomScanNodeProperties : ScanNodeProperties
     [ConfigSection("PingScan")]
     [ConfigDescription("Adds a scan node to every item that didn't already have one.")]
     public static ConfigData<bool> EnableScanNodeForTools = new(true);
+    
+    public bool IsBatteryUpdate { get; private set; }
 
     private void Awake()
     {
@@ -14,6 +16,7 @@ public class CustomScanNodeProperties : ScanNodeProperties
         
         if (grabbableObject != null && grabbableObject.itemProperties.requiresBattery)
         {
+            IsBatteryUpdate = true;
             gameObject.AddComponent<BatteryTextUpdater>();
         }
     }