+++ Plugin.cs
@@ -4,23 +4,31 @@ using UnityMDK.Config;
 
 namespace ScanTweaks;
 
-[BepInDependency(LethalMDK.LethalMDK.ModGuid)]
-[BepInDependency(UnityMDK.UnityMDK.ModGuid)]
-[BepInPlugin(ScanTweaks.ModGuid, ScanTweaks.ModName, ScanTweaks.ModVersion)]
+[BepInDependency(PluginInfo.ReservedItemSlotsCoreGuid, BepInDependency.DependencyFlags.SoftDependency)]
+[BepInDependency(LethalMDK.LethalMDK.ModGuid, BepInDependency.DependencyFlags.HardDependency)]
+[BepInDependency(UnityMDK.UnityMDK.ModGuid, BepInDependency.DependencyFlags.HardDependency)]
+[BepInPlugin(PluginInfo.ModGuid, PluginInfo.ModName, PluginInfo.ModVersion)]
 public class PluginInitializer : BaseUnityPlugin
 {
-    private readonly Harmony _harmonyInstance = new(ScanTweaks.ModGuid);
+    private readonly Harmony _harmonyInstance = new(PluginInfo.ModGuid);
         
     private void Awake()
     {
         ConfigBinder.BindAll(Config);
         _harmonyInstance.PatchAll();
+
+        if (BepInEx.Bootstrap.Chainloader.PluginInfos.ContainsKey(PluginInfo.ReservedItemSlotsCoreGuid))
+        {
+            PingScan.SetPocketedGrabbableComparer(new ReservedPocketedGrabbableComparer());
+        }
     }
 }
 
-public static class ScanTweaks
+public static class PluginInfo
 {
     public const string ModGuid = "Saradora.ScanTweaks";
-    public const string ModVersion = "1.7.0";
+    public const string ModVersion = "1.8.0";
     public const string ModName = "Scan Tweaks";
+
+    public const string ReservedItemSlotsCoreGuid = "FlipMods.ReservedItemSlotCore";
 }